requirements.txt
absl-py==1.2.0
aiohttp==3.8.3
aiosignal==1.2.0
altair==4.2.0
apps==0.0.0.0.dev20200918
asgiref==3.5.2
asttokens==2.0.8
astunparse==1.6.3
async-timeout==4.0.2
attrs==22.1.0
autopep8==1.7.0
backcall==0.2.0
black==22.8.0
blinker==1.5
Brotli==1.0.9
cachetools==5.2.0
certifi==2022.6.15.1
chardet==3.0.4
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
commonmark==0.9.1
contourpy==1.0.5
convert-html-to-dash==0.1.2
cv==1.0.0
cycler==0.11.0
dash==2.6.2
dash-bootstrap-components==1.2.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
debugpy==1.6.3
decorator==5.1.1
discord==2.0.0
discord.py==2.0.1
Django==4.1.1
entrypoints==0.4
executing==1.0.0
Flask==2.2.2
Flask-Compress==1.13
flatbuffers==2.0.7
fonttools==4.37.3
frozenlist==1.3.1
gast==0.4.0
geojson==2.5.0
gitdb==4.0.9
GitPython==3.1.27
gmaps==0.9.0
google-auth==2.11.1
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googletrans==3.1.0a0
grpcio==1.49.1
h11==0.9.0
h2==3.2.0
h5py==3.7.0
hpack==3.0.0
hstspreload==2022.9.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
importlib-metadata==4.12.0
ipykernel==6.15.3
ipython==8.5.0
ipywidgets==8.0.2
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.16.0
jupyter-core==4.11.1
jupyter_client==7.3.5
jupyterlab-widgets==3.0.3
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
keras-resnet==0.2.0
keras-utils==1.0.13
kiwisolver==1.4.4
libclang==14.0.6
lxml==4.9.1
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.6.0
matplotlib-inline==0.1.6
MultiApp==1.0
multidict==6.0.2
mypy-extensions==0.4.3
nest-asyncio==1.5.5
numpy==1.23.3
oauthlib==3.2.1
opencv-python==4.6.0.66
opt-einsum==3.3.0
packaging==21.3
pandas==1.5.0
parso==0.8.3
pathspec==0.10.1
pickleshare==0.7.5
Pillow==9.2.0
platformdirs==2.5.2
plotly==5.10.0
prompt-toolkit==3.0.31
protobuf==3.19.5
psutil==5.9.2
pure-eval==0.2.2
pyarrow==9.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.9.1
pydeck==0.8.0b3
Pygments==2.13.0
Pympler==1.0.1
pyparsing==3.0.9
pyrsistent==0.18.1
python-dateutil==2.8.2
pytz==2022.2.1
pytz-deprecation-shim==0.1.0.post0
pywin32==304
PyYAML==6.0
pyzmq==24.0.1
requests==2.28.1
requests-oauthlib==1.3.1
rfc3986==1.5.0
rich==12.5.1
rsa==4.9
scikit-learn==1.1.2
scipy==1.9.1
semver==2.13.0
six==1.16.0
sklearn==0.0
smmap==5.0.0
sniffio==1.3.0
sqlparse==0.4.2
stack-data==0.5.0
streamlit==1.13.0
streamlit-lottie==0.0.3
termcolor==2.0.1
threadpoolctl==3.1.0
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
tornado==6.2
tqdm==4.64.1
traitlets==5.4.0
typing_extensions==4.3.0
tzdata==2022.2
tzlocal==4.2
urllib3==1.26.12
validators==0.20.0
watchdog==2.1.9
wcwidth==0.2.5
Werkzeug==2.2.2
widgetsnbextension==4.0.3
wrapt==1.14.1
yarl==1.8.1
zipp==3.8.1
